cryptohack/requirements.txt
pycryptodome==3.21.0
pwntools==4.13.1
